feat: implement deprecation of legacy request methods
