+++ tests/main/tests/integration/application-test.js
@@ -2,9 +2,9 @@
 import Controller from '@ember/controller';
 import * as s from '@ember/service';
 
+import Store from 'main-test-app/services/store';
 import { module, test } from 'qunit';
 
-import Store from 'ember-data/store';
 import { setupTest } from 'ember-qunit';
 
 import JSONAPIAdapter from '@ember-data/adapter/json-api';
