+++ tests/main/tests/integration/cache/spec-cache-errors-test.ts
@@ -1,8 +1,8 @@
 import EmberObject from '@ember/object';
 
+import Store from 'main-test-app/services/store';
 import { module, test } from 'qunit';
 
-import Store from 'ember-data/store';
 import { setupTest } from 'ember-qunit';
 
 import { InvalidError } from '@ember-data/adapter/error';
