+++ tests/main/tests/integration/multiple-stores-test.js
@@ -1,8 +1,8 @@
 // we intentionally test against the ember-data version here
 // because the ember-data/store uses DefaultRecordData while @ember-data/store does not
+import Store from 'main-test-app/services/store';
 import { module, test } from 'qunit';
 
-import Store from 'ember-data/store';
 import { setupTest } from 'ember-qunit';
 
 import Adapter from '@ember-data/adapter';
