+++ tests/main/tests/integration/store/serializer-for-test.js
@@ -1,8 +1,8 @@
 import { settled } from '@ember/test-helpers';
 
+import Store from 'main-test-app/services/store';
 import { module, test } from 'qunit';
 
-import Store from 'ember-data/store';
 import { setupTest } from 'ember-qunit';
 
 class TestAdapter {
