+++ tests/main/tests/test-helper.js
@@ -11,10 +11,16 @@ import start from 'ember-exam/test-support/start';
 
 import { setBuildURLConfig } from '@ember-data/request-utils';
 import configureAsserts from '@ember-data/unpublished-test-infra/test-support/asserts/index';
+import { Store } from '@warp-drive/core';
 import { setConfig, setTestId } from '@warp-drive/holodeck';
+import { Model, restoreDeprecatedModelRequestBehaviors } from '@warp-drive/legacy/model';
+import { restoreDeprecatedStoreBehaviors } from '@warp-drive/legacy/store';
 
 import Application from '../app';
 import config from '../config/environment';
+
+restoreDeprecatedStoreBehaviors(Store);
+restoreDeprecatedModelRequestBehaviors(Model);
 
 const MockHost = `https://${window.location.hostname}:${Number(window.location.port) + 1}`;
 setBuildURLConfig({
