+++ tests/main/tests/unit/store/asserts-test.js
@@ -1,8 +1,8 @@
 import { settled } from '@ember/test-helpers';
 
+import Store from 'main-test-app/services/store';
 import { module } from 'qunit';
 
-import Store from 'ember-data/store';
 import { setupTest } from 'ember-qunit';
 
 import Model from '@ember-data/model';
