+++ warp-drive-packages/build-config/src/deprecation-versions.ts
@@ -28,3 +28,4 @@ export const ENABLE_LEGACY_SCHEMA_SERVICE = '5.4';
 export const DEPRECATE_EMBER_INFLECTOR = '5.3';
 export const DISABLE_7X_DEPRECATIONS = '7.0';
 export const DEPRECATE_TRACKING_PACKAGE = '5.5';
+export const ENABLE_LEGACY_REQUEST_METHODS = '5.6';
