+++ warp-drive-packages/core/src/types.ts
@@ -7,7 +7,8 @@
 export type { StableRecordIdentifier, ResourceKey } from './types/identifier.ts';
 
 export type { CacheCapabilitiesManager } from './store/-types/q/cache-capabilities-manager.ts';
-export type { ModelSchema } from './store/-types/q/ds-model.ts';
+// FIXME this should come from somewhere more intelligent
+export type { ModelSchema } from './store/deprecated/-private.ts';
 export type { SchemaService } from './store/-types/q/schema-service.ts';
 export type {
   BaseFinderOptions,
